// src/types.ts
export interface DepartmentGroupProps {
  departmentGroupCode: string;
  departmentGroupName: string;
  id: number;
  isActive: number;
}